Cleanup single project view to its own component
